+++ menu/market.go
@@ -1263,7 +1263,7 @@ func RunNFAMarket(d *dreams.AppObject, cont *fyne.Container) {
 
 // Get search filters from on chain store
 func GetFilters(check string) (filter []string) {
-	if stored, ok := rpc.GetStringKey(rpc.RatingSCID, check, rpc.Daemon.Rpc).(string); ok {
+	if stored, ok := rpc.GetStringKey(rpc.RatingSCID, check, rpc.Daemon.Endpoint).(string); ok {
 		if h, err := hex.DecodeString(stored); err == nil {
 			if err = json.Unmarshal(h, &filter); err != nil {
 				logger.Errorln("[GetFilters]", check, err)